Updated Knight CheckMove to remove error
When loading the private int variable numArrayPos with the variable
numPosition I was taking away an additional 1 from the variable which
then caused the arrays to be out of line with each other. 

When checking the first if statement of each move numArrayPos was being
read as -2 rather than -1. This caused the mouse-clicks to pass the
criteria for north checks when it shouldn't have.