Screen.java Update multiple changes.
Removed several methods that were associated with mouse controls from
the Screen class and placed them within their own class "MouseControl."

Once the methods were removed and MouseControl was optimised to work in
conjuction with Screen.java I added a MouseControl reference and object
to hand all Mouse input within the Screen class.

With the removal of the input() method and the variables I had to update
the showLegalMove() method with getters from the MouseControl class.