Added Controls Package and MouseControls Class
Moved the input() method as well as several other methods involved in
working out mouse control to a seperate class called MouseControl inside
a new Package called Controls.

input() has been updated to accept method variables for;
Screen Size Height, which is used with Mouse.getY as the OpenGL Y-Axis
is inverted. 
BoardPosition 3D Array needs to be fed into the method for everything to
work.
ClickOutline 2D Array which would be used to draw the blue and green
squares which show mouse click positions!

Two getters were also added to the class so ChaX and IntY, which are
private variables, could be accessed outside the class.